backend/requirements.txt
Flask
gunicorn
bitcoinlib
python-dotenv
cryptos